update changes for version 2
