+++ NAMESPACE
@@ -29,10 +29,8 @@ export(falseOmissionRate)
 export(falsePositiveRate)
 export(fitCNNTorch)
 export(fitGLMModel)
-export(fitMLPTorch)
 export(fitPlp)
 export(fitRNNTorch)
-export(fitLRTorch)
 export(getAttritionTable)
 export(getCalibration)
 export(getCovariateData)
@@ -102,7 +100,6 @@ import(Cyclops)
 import(DatabaseConnector)
 import(FeatureExtraction)
 import(bit)
-import(plotly)
 importFrom(Rcpp,evalCpp)
 importFrom(SqlRender,loadRenderTranslateSql)
 importFrom(SqlRender,translateSql)
