+++ R/CNNTorch.R
@@ -20,8 +20,9 @@
 #' @param nbfilters  The number of filters
 #' @param epochs     The number of epochs
 #' @param seed       A seed for the model
-#' @param time_window       A time window to get temporal data
-#' @param class_weight      weight the class in imblanced data
+#' @param class_weight   The class weight used for imbalanced data: 
+#'                           0: Inverse ratio between positives and negatives
+#'                          -1: Focal loss
 #' @param cnn_type      It can be normal 'CNN', 'CNN_MLF' with multiple kernels with different kernel size, 'CNN_MIX' and 'CNN_MULTI'
 #' 
 #' @examples
@@ -29,7 +30,7 @@
 #' model.cnnTorch <- setCNNTorch()
 #' }
 #' @export
-setCNNTorch <- function(nbfilters=c(16, 32), epochs=c(20, 50), seed=0, time_window = 12, class_weight = 0, cnn_type = 'CNN'){
+setCNNTorch <- function(nbfilters=c(16, 32), epochs=c(20, 50), seed=0, class_weight = 0, cnn_type = 'CNN'){
   
   # test python is available and the required dependancies are there:
   if (!PythonInR::pyIsConnected()){
@@ -40,10 +41,11 @@ setCNNTorch <- function(nbfilters=c(16, 32), epochs=c(20, 50), seed=0, time_wind
     }  
     )
   }
-  result <- list(model='fitCNNTorch', param= expand.grid(nbfilters=nbfilters,
-                                                         epochs=epochs, seed=ifelse(is.null(seed),'NULL', seed), 
-											time_window = time_window, class_weight = class_weight, cnn_type = cnn_type),
-                 name='CNN Torch')
+  result <- list(model='fitCNNTorch', param=split(expand.grid(nbfilters=nbfilters,
+                                            epochs=epochs, seed=ifelse(is.null(seed),'NULL', seed), 
+											class_weight = class_weight, cnn_type = cnn_type),
+											1:(length(nbfilters)*length(epochs)) ),
+                 					  name='CNN Torch')
   
   class(result) <- 'modelSettings' 
   
@@ -64,10 +66,9 @@ fitCNNTorch <- function(population, plpData, param, search='grid', quiet=F,
   }
   
   # connect to python if not connected
-  if ( !PythonInR::pyIsConnected() ){ 
+  if ( !PythonInR::pyIsConnected() || .Platform$OS.type=="unix"){ 
     PythonInR::pyConnect()
   }
-
   # return error if we can't connect to python
   if ( !PythonInR::pyIsConnected() )
     stop('Python not connect error')
@@ -90,56 +91,30 @@ fitCNNTorch <- function(population, plpData, param, search='grid', quiet=F,
   covariates$rowIdPython <- covariates$rowId -1 #to account for python/r index difference
   PythonInR::pySet('covariates', as.matrix(covariates[,c('rowIdPython','covariateId','timeId', 'covariateValue')]))
   
-  covariateRef <- ff::as.ram(plpData$covariateRef)
-  inc <- 1:ncol(covariateRef)  
   # save the model to outLoc  TODO: make this an input or temp location?
   outLoc <- file.path(getwd(),'python_models')
   # clear the existing model pickles
   for(file in dir(outLoc))
     file.remove(file.path(outLoc,file))
-  
-  covariateRef <- ff::as.ram(plpData$covariateRef)
-  incs <- rep(1, nrow(covariateRef))
-  covariateRef$included <- incs
-  #covariateRef$value <- unlist(varImp)
-  all_auc <- c()
-  
-  for(i in 1:nrow(param)){
-    
-    # do inc-1 to go to python index as python starts at 0, R starts at 1
-    PythonInR::pySet('included', as.matrix(inc-1), 
-                     namespace = "__main__", useNumpy = TRUE)
-    
-    # then run standard python code
-    auc <- do.call(trainCNNTorch,list(epochs=as.character(param$epochs[i]), nbfilters = as.character(param$nbfilters[i]), 
-                                      seed = as.character(param$seed[i]), time_window = as.character(param$time_window[i]), 
-									  class_weight = as.character(param$class_weight[i]), cnn_type = as.character(param$cnn_type[i]),
-									  train = TRUE))
-    
-    all_auc <- c(all_auc, auc)
-    writeLines(paste0('Model with settings: epochs: ',param$epochs[i], 
-                      'nbfilters: ', param$nbfilters[i], 'seed: ', param$seed[i], 'time_window:', param$time_window[i], ' obtained AUC of ', auc))
-  }
-  
-  hyperSummary <- cbind(param, cv_auc=all_auc)
-  
-  # run model:
+
   outLoc <- file.path(getwd(),'python_models')
   PythonInR::pySet("modelOutput",outLoc)
+
+  # do cross validation to find hyperParameter
+  hyperParamSel <- lapply(param, function(x) do.call(trainCNNTorch, c(x, train=TRUE)  ))
+ 
+  hyperSummary <- cbind(do.call(rbind, param), unlist(hyperParamSel))
   
-  
-  # ToDo: I do not like this list creation
-  finalModel <- do.call(trainCNNTorch,list(epochs=as.character(param$epochs[which.max(all_auc)]), 
-                                           nbfilters=as.character(param$nbfilters[which.max(all_auc)]), 
-                                           seed = as.character(param$seed[which.max(all_auc)]), 
-                                           time_window = as.character(param$time_window[which.max(all_auc)]),
-										                       class_weight = as.character(param$class_weight[which.max(all_auc)]),
-										                       cnn_type = as.character(param$cnn_type[which.max(all_auc)]),
-                                           train = FALSE))
-  
+  #now train the final model and return coef
+  bestInd <- which.max(abs(unlist(hyperParamSel)-0.5))[1]
+  finalModel <- do.call(trainCNNTorch, c(param[[bestInd]], train=FALSE))
+
+  covariateRef <- ff::as.ram(plpData$covariateRef)
+  incs <- rep(1, nrow(covariateRef)) 
+  covariateRef$included <- incs
   
   modelTrained <- file.path(outLoc) 
-  param.best <- NULL
+  param.best <- param[[bestInd]]
   
   comp <- start-Sys.time()
   
@@ -164,12 +139,12 @@ fitCNNTorch <- function(population, plpData, param, search='grid', quiet=F,
 }
 
 
-trainCNNTorch <- function(epochs=50, nbfilters = 16, seed=0, time_window = 12, class_weight= 0, cnn_type = 'CNN', train=TRUE){
+trainCNNTorch <- function(epochs=50, nbfilters = 16, seed=0, class_weight= 0, cnn_type = 'CNN', train=TRUE){
   #PythonInR::pyExec(paste0("size = ",size))
   PythonInR::pyExec(paste0("epochs = ",epochs))
   PythonInR::pyExec(paste0("nbfilters = ",nbfilters))
   PythonInR::pyExec(paste0("seed = ",seed))
-  PythonInR::pyExec(paste0("time_window = ",time_window))
+  #PythonInR::pyExec(paste0("time_window = ",time_window))
   PythonInR::pyExec(paste0("class_weight = ",class_weight))
   if (cnn_type == 'CNN'){
     PythonInR::pyExec("model_type = 'CNN'")
@@ -186,7 +161,8 @@ trainCNNTorch <- function(epochs=50, nbfilters = 16, seed=0, time_window = 12, c
     PythonInR::pyExec("train = True")
   if(!train)
     PythonInR::pyExec("train = False")
-  
+  python_dir <- system.file(package='PatientLevelPrediction','python')
+  PythonInR::pySet("python_dir", python_dir)
   # then run standard python code #learningcurve.py #deepTorch.py
   PythonInR::pyExecfile(system.file(package='PatientLevelPrediction','python','deepTorch.py'))
