+++ R/MLPTorch.R
@@ -17,18 +17,21 @@
 # limitations under the License.
 
 #' Create setting for neural network model with python 
-#' @param size       The number of hidden nodes
-#' @param w_decay      The l2 regularisation
-#' @param seed       A seed for the model
-#' @param epochs     The number of epochs 
-#' @param class_weight   weight the class in imblanced data
+#' @param size           The number of hidden nodes
+#' @param w_decay        The l2 regularisation
+#' @param seed           A seed for the model
+#' @param epochs         The number of epochs 
+#' @param class_weight   The class weight used for imbalanced data: 
+#'                           0: Inverse ratio between positives and negatives
+#'                          -1: Focal loss
 #'
 #' @examples
 #' \dontrun{
 #' model.mlpTorch <- setMLPTorch()
 #' }
 #' @export
-setMLPTorch <- function(size=c(500, 1000), w_decay=c(0.0005, 0.005), epochs=c(20, 50), seed=0, class_weight = 0){
+setMLPTorch <- function(size=c(500, 1000), w_decay=c(0.0005, 0.005), 
+                        epochs=c(20, 50), seed=0, class_weight = 0){
   
   # test python is available and the required dependancies are there:
   if (!PythonInR::pyIsConnected()){
@@ -39,20 +42,16 @@ setMLPTorch <- function(size=c(500, 1000), w_decay=c(0.0005, 0.005), epochs=c(20
     }  
     )
   }
-  result <- list(model='fitMLPTorch', param= expand.grid(size=size, w_decay=w_decay,
-                 epochs=epochs, seed=ifelse(is.null(seed),'NULL', seed), class_weight = ifelse(is.null(class_weight),'NULL', class_weight)), 
-                 name='MLP Torch')
-  
-  #result <- list(model='fitMLPTorch', 
-  #               param= c(size,epochs,seed),
-  #               name='MLP Torch')
+  result <- list(model='fitMLPTorch', param=split(expand.grid(size=size, w_decay=w_decay,
+                                            epochs=epochs, seed=ifelse(is.null(seed),'NULL', seed), class_weight = class_weight),
+				                            1:(length(size)*length(w_decay)*length(epochs)) ),
+                                      name='MLP Torch')
   
   class(result) <- 'modelSettings' 
   
   return(result)
 }
 
-#' @export
 fitMLPTorch <- function(population, plpData, param, search='grid', quiet=F,
                    outcomeId, cohortId, ...){
   
@@ -60,23 +59,23 @@ fitMLPTorch <- function(population, plpData, param, search='grid', quiet=F,
   if(!'ffdf'%in%class(plpData$covariates))
     stop('Needs plpData')
   
+  # check population has indexes column, which is used to split training with different folds and testing set
   if(colnames(population)[ncol(population)]!='indexes'){
     warning('indexes column not present as last column - setting all index to 1')
     population$indexes <- rep(1, nrow(population))
   }
   
   # connect to python if not connected
-  if ( !PythonInR::pyIsConnected() ){ 
+  if ( !PythonInR::pyIsConnected() || .Platform$OS.type=="unix"){ 
     PythonInR::pyConnect()
   }
-  
   
   # return error if we can't connect to python
   if ( !PythonInR::pyIsConnected() )
     stop('Python not connect error')
   
-  PythonInR::pyOptions("numpyAlias", "np")
-  PythonInR::pyOptions("useNumpy", TRUE)
+  PythonInR::pyOptions("numpyAlias", "np")  #NEEDED?
+  PythonInR::pyOptions("useNumpy", TRUE) #NEEDED?
   PythonInR::pyImport("numpy", as='np')
   
   start <- Sys.time()
@@ -86,69 +85,36 @@ fitMLPTorch <- function(population, plpData, param, search='grid', quiet=F,
   
   # convert plpData in coo to python:
   x <- toSparsePython(plpData,population, map=NULL)
-  covariateRef <- ff::as.ram(plpData$covariateRef)
-  inc <- 1:ncol(covariateRef)  
+
+  #inc <- 1:ncol(covariateRef)  
   # save the model to outLoc  TODO: make this an input or temp location?
   outLoc <- file.path(getwd(),'python_models')
   # clear the existing model pickles
   for(file in dir(outLoc))
     file.remove(file.path(outLoc,file))
 
-  covariateRef <- ff::as.ram(plpData$covariateRef)
-  incs <- rep(1, nrow(covariateRef))
-  covariateRef$included <- incs
   #covariateRef$value <- unlist(varImp)
-  all_auc <- c()
-  
-  for(i in 1:nrow(param)){
-    # set variable params - do loop  
-    #PythonInR::pyExec(paste0("size = int(",param$size[i],")"))
-    #PythonInR::pyExec(paste0("w_decay = int(",param$w_decay[i],")"))
-    #PythonInR::pyExec(paste0("epochs = int(",param$epochs[i],")"))
-    #PythonInR::pySet("epochs",param$epochs[i])
-    #PythonInR::pySet("train", FALSE)
-    ##PythonInR::pySet("dataLocation" ,plpData$covariates)
     
-    # do inc-1 to go to python index as python starts at 0, R starts at 1
-    ##PythonInR::pyImport("numpy", as="np")
-    PythonInR::pySet('included', as.matrix(inc-1), 
-                     namespace = "__main__", useNumpy = TRUE)
-    
-    #mapping = sys.argv[5] # this contains column selection/ordering 
-    #missing = sys.argv[6] # this contains missing
-    
-    # then run standard python code
-    #PythonInR::pyExecfile(system.file(package='PatientLevelPrediction','python','mlp_torch.py'))
-    auc <- do.call(trainMLPTorch,list(size=as.character(param$size[i]), epochs=as.character(param$epochs[i]), w_decay = as.character(param$w_decay[i]), 
-                                                        seed = as.character(param$seed[i]), class_weight = as.character(param$class_weight[i]), train = TRUE))
-
-    # close python
-    
-    ##pred <- read.csv(file.path(outLoc,i,'prediction.txt'), header=F)
-    ##colnames(pred) <- c('rowId','outcomeCount','indexes', 'value')
-    ##auc <- PatientLevelPrediction::computeAuc(pred)
-    all_auc <- c(all_auc, auc)
-    writeLines(paste0('Model with settings: size:',param$size[i],' epochs: ',param$epochs[i], 
-                      'w_decay: ', param$w_decay[i], 'seed: ', param$seed[i], 'class_weight: ', param$class_weight[i], ' obtained AUC of ', auc))
-  }
-  
-  hyperSummary <- cbind(param, cv_auc=all_auc)
-  
   # run model:
   outLoc <- file.path(getwd(),'python_models')
   PythonInR::pySet("modelOutput",outLoc)
 
+  #do cross validation to find hyperParameter
+  hyperParamSel <- lapply(param, function(x) do.call(trainMLPTorch, c(x, train=TRUE)  ))
+
   
-  # ToDo: I do not like this list creation
-  finalModel <- do.call(trainMLPTorch,list(size=as.character(param$size[which.max(all_auc)]), 
-                                           epochs=as.character(param$epochs[which.max(all_auc)]), 
-                                           w_decay=as.character(param$w_decay[which.max(all_auc)]), 
-                                           seed = as.character(param$seed[which.max(all_auc)]), 
-                                           class_weight = as.character(param$class_weight[which.max(all_auc)]), train = FALSE))
+  hyperSummary <- cbind(do.call(rbind, param), unlist(hyperParamSel))
   
+  #now train the final model
+  bestInd <- which.max(abs(unlist(hyperParamSel)-0.5))[1]
+  finalModel <- do.call(trainMLPTorch, c(param[[bestInd]], train=FALSE))
+
+  covariateRef <- ff::as.ram(plpData$covariateRef)
+  incs <- rep(1, nrow(covariateRef)) 
+  covariateRef$included <- incs
   
   modelTrained <- file.path(outLoc) 
-  param.best <- NULL
+  param.best <- param[[bestInd]]
   
   comp <- start-Sys.time()
   
@@ -180,6 +146,8 @@ trainMLPTorch <- function(size=200, epochs=100, w_decay = 0.001, seed=0, class_w
   PythonInR::pyExec(paste0("seed = ",seed))
   PythonInR::pyExec(paste0("class_weight = ",class_weight))
   PythonInR::pyExec("model_type = 'MLP'")
+  python_dir <- system.file(package='PatientLevelPrediction','python')
+  PythonInR::pySet("python_dir", python_dir)
   if(train)
     PythonInR::pyExec("train = True")
   if(!train)
