+++ man/runPlp.Rd
@@ -4,16 +4,18 @@
 \alias{runPlp}
 \title{runPlp - Train and evaluate the model}
 \usage{
-runPlp(population, plpData, modelSettings, testSplit = "time",
-  testFraction = 0.25, splitSeed = NULL, nfold = 3, indexes = NULL,
-  save = NULL, saveModel = T, verbosity = futile.logger::INFO,
-  timeStamp = FALSE, analysisId = NULL)
+runPlp(population, plpData, minCovariateFraction = 0.001, modelSettings,
+  testSplit = "time", testFraction = 0.25, splitSeed = NULL, nfold = 3,
+  indexes = NULL, save = NULL, saveModel = T,
+  verbosity = futile.logger::INFO, timeStamp = FALSE, analysisId = NULL)
 }
 \arguments{
 \item{population}{The population created using createStudyPopulation() who will be used to develop the model}
 
 \item{plpData}{An object of type \code{plpData} - the patient level prediction
 data extracted from the CDM.}
+
+\item{minCovariateFraction}{The minimum fraction of target population who must have a covariate for it to be included in the model training}
 
 \item{modelSettings}{An object of class \code{modelSettings} created using one of the function:
 \itemize{
@@ -109,7 +111,7 @@ population <- createStudyPopulation(plpData,outcomeId=2042,
 #and results are saved to file.path('C:','User','home')
 model.lr <- lassoLogisticRegression.set()
 mod.lr <- runPlp(population=population,
-                        plpData= plpData,
+                        plpData= plpData, minCovariateFraction = 0.001,
                         modelSettings = model.lr ,
                         testSplit = 'time', testFraction=0.3, 
                         nfold=3, indexes=NULL,
