+++ man/setCNNTorch.Rd
@@ -5,7 +5,7 @@
 \title{Create setting for CNN model with python}
 \usage{
 setCNNTorch(nbfilters = c(16, 32), epochs = c(20, 50), seed = 0,
-  time_window = 12, class_weight = 0, cnn_type = "CNN")
+  class_weight = 0, cnn_type = "CNN")
 }
 \arguments{
 \item{nbfilters}{The number of filters}
@@ -14,9 +14,9 @@ setCNNTorch(nbfilters = c(16, 32), epochs = c(20, 50), seed = 0,
 
 \item{seed}{A seed for the model}
 
-\item{time_window}{A time window to get temporal data}
-
-\item{class_weight}{weight the class in imblanced data}
+\item{class_weight}{The class weight used for imbalanced data: 
+ 0: Inverse ratio between positives and negatives
+-1: Focal loss}
 
 \item{cnn_type}{It can be normal 'CNN', 'CNN_MLF' with multiple kernels with different kernel size, 'CNN_MIX' and 'CNN_MULTI'}
 }
